+++ GPTifier/CMakeLists.txt
@@ -80,7 +80,6 @@ add_executable(
   "gpt"
   "src/configs.cpp"
   "src/datadir.cpp"
-  "src/help_messages.cpp"
   "src/interface/command_chats.cpp"
   "src/interface/command_costs.cpp"
   "src/interface/command_embed.cpp"
@@ -90,6 +89,7 @@ add_executable(
   "src/interface/command_run.cpp"
   "src/interface/command_short.cpp"
   "src/interface/command_test.cpp"
+  "src/interface/help_messages.cpp"
   "src/interface/params.cpp"
   "src/main.cpp"
   "src/networking/api_openai_admin.cpp"
