+++ GPTifier/include/interface/help_messages.hpp
