+++ GPTifier/src/interface/command_chats.cpp
@@ -1,6 +1,6 @@
 #include "interface/command_chats.hpp"
 
-#include "help_messages.hpp"
+#include "interface/help_messages.hpp"
 #include "interface/params.hpp"
 #include "serialization/chat_completions.hpp"
 #include "utils.hpp"
