+++ GPTifier/src/interface/command_embed.cpp
@@ -2,7 +2,7 @@
 
 #include "configs.hpp"
 #include "datadir.hpp"
-#include "help_messages.hpp"
+#include "interface/help_messages.hpp"
 #include "interface/params.hpp"
 #include "serialization/embeddings.hpp"
 #include "utils.hpp"
