+++ GPTifier/src/interface/command_short.cpp
@@ -1,6 +1,6 @@
 #include "interface/command_short.hpp"
 
-#include "help_messages.hpp"
+#include "interface/help_messages.hpp"
 #include "interface/params.hpp"
 #include "selectors.hpp"
 #include "serialization/chat_completions.hpp"
