+++ GPTifier/src/interface/help_messages.cpp
@@ -1,4 +1,4 @@
-#include "help_messages.hpp"
+#include "interface/help_messages.hpp"
 
 #include "utils.hpp"
