+++ GPTifier/src/main.cpp
@@ -1,5 +1,4 @@
 #include "configs.hpp"
-#include "help_messages.hpp"
 #include "interface/command_chats.hpp"
 #include "interface/command_costs.hpp"
 #include "interface/command_embed.hpp"
@@ -9,6 +8,7 @@
 #include "interface/command_run.hpp"
 #include "interface/command_short.hpp"
 #include "interface/command_test.hpp"
+#include "interface/help_messages.hpp"
 
 #include <fmt/core.h>
 #include <json.hpp>
